Work around OSVERSION checks, fix sysctl for number of cpus
The makefiles use OSVERSION extensively to check for FreeBSD features
to either abort the build or do something differently.  Inserting the
DragonFly OSVERSION there is meaningless.

Hardcode OSVERSION to 9999999 to simulate the latest FreeBSD version
and hopefully this will allow builds on DragonFly by default.  If
not, a patch will have to be used to remove the check as
Makefile.DragonFly trick doesn't work for all targets.