Fix segfault when stopping chromium.
The TimeZoneMonitorLinuxImpl constructor (running on  the UI thread)
triggers a function to run on the FILE thread with PostTask. If that call
runs before the constructor is finished, it seems to destroy the
TimeZoneMonitorLinuxImpl object (because ref counting doesn't seem to be
correctly initialized yet, while the constructor isn't finished).
When chromium is terminated later, we try to destroy the already destroyed
TimeZoneMonitor causing a segfault.
I observed this problem on DragonFly, so I'm not sure if this actually
happens on FreeBSD.